requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
Django==4.2.7
executing==2.0.0
fastapi==0.108.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.2
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kubernetes==29.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pdftotext==2.2.2
pickleshare==0.7.5
platformdirs==3.11.0
posthog==3.1.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
psycopg==3.1.12
psycopg-binary==3.1.12
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.36.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.25.11
uvicorn==0.25.0
watchfiles==0.21.0
wcwidth==0.2.8
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
